[OPMONDEV-176]: move test infra code to test_utils.py
